noticeDetail 수정정
