+++ src/pages/Notice/NoticeDetail.css
@@ -43,3 +43,12 @@
     height: 1px;
     background-color: #adadad;
 }
+
+.notice_detail_listbtn{
+    float: right;
+    width: 100px; height: 30px;
+    margin: 10px 0;
+    background-color: white;
+    border: 1px solid #adadad;
+    border-radius: 5px;
+}
