+++ src/pages/Party/PartyDetail.css
@@ -143,43 +143,6 @@
     right: 16px;
 }
 
-.party_datail_reply_item{
-    margin-top: 30px;
-    display: flex;
-    gap: 38px;
-}
-
-.party_datail_reply_item > .image_container{
-    margin-top: 8px;
-    flex-shrink: 0;
-    width: 50px; height: 50px;
-    border-radius: 50px;
-    overflow: hidden;
-}
-
-.party_datail_reply_item > .image_container > img{
-    display: block;
-    width: 100%; height: 100%;
-    object-fit: cover;
-    object-position: center;
-}
-
-.party_datail_reply_item .insight {
-    display: flex;
-    gap: 10px;
-    color: #aeaeae;
-}
-
-.party_datail_reply_item .insight > span:first-child{
-    color: #606060;
-}
-
-.party_datail_reply_item p{
-    white-space: pre-line;
-    margin: 0;
-    margin-top: 8px;;
-}
-
 .party_detail_divider{
     margin-top: 30px;
     width: 100%; height: 1px;
