+++ src/pages/Recipe/RecipeDetail.css
@@ -221,7 +221,7 @@
   right: 16px;
 }
 
-.recipe_detail_reply_item {
+/* .recipe_detail_reply_item {
   margin-top: 30px;
   display: flex;
   gap: 38px;
@@ -258,7 +258,7 @@
   white-space: pre-line;
   margin: 0;
   margin-top: 8px;
-}
+} */
 
 .recipe_detail_divider {
   width: 100%;
